% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{generate_sequel}
\alias{generate_sequel}
\title{Generate a sequel name}
\usage{
generate_sequel(x, n)
}
\arguments{
\item{x}{numeric eg. 3}

\item{n}{number of outputs}
}
\value{
sequel with rhyming end
}
\description{
Generate a sequel name
}
