% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{get_latest_file}
\alias{get_latest_file}
\title{Get the name of the latest version of a file or directory}
\usage{
get_latest_file(pattern)
}
\arguments{
\item{pattern}{Add the pattern that the file or directory should start with.
If the file is in another directory this should also have the file path}
}
\value{
latest_file_name
}
\description{
Get the name of the latest version of a file or directory
}
