% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{get_latest_file}
\alias{get_latest_file}
\title{Get the name of the latest version of a file or directory}
\usage{
get_latest_file(path = NA, pattern)
}
\arguments{
\item{path}{path to files, defaults to wd}

\item{pattern}{Add the pattern that the file or directory should start with.}
}
\value{
latest_file_name
}
\description{
Get the name of the latest version of a file or directory
}
